fix tests

+++ test/e2e/skipinboundtags/e2e_suite_test.go
@@ -17,19 +17,26 @@ func TestE2E(t *testing.T) {
 	test.RunE2ESpecs(t, "E2E Skip Inbound Tags Kubernetes Suite")
 }
 
-var _ = framework.E2EBeforeSuite(func() {
-	skipinboundtags.KubeCluster = framework.NewK8sCluster(framework.NewTestingT(), framework.Kuma1, framework.Silent)
+var _ = framework.E2ESynchronizedBeforeSuite(
+	func() []byte {
+		skipinboundtags.KubeCluster = framework.NewK8sCluster(framework.NewTestingT(), framework.Kuma1, framework.Silent)
 
-	err := framework.NewClusterSetup().
-		Install(framework.Kuma(config_core.Zone,
-			framework.WithEnv("KUMA_EXPERIMENTAL_INBOUND_TAGS_DISABLED", "true"),
-		)).
-		Setup(skipinboundtags.KubeCluster)
+		err := framework.NewClusterSetup().
+			Install(framework.Kuma(config_core.Zone,
+				framework.WithEnv("KUMA_EXPERIMENTAL_INBOUND_TAGS_DISABLED", "true"),
+			)).
+			Setup(skipinboundtags.KubeCluster)
 
-	Expect(err).ToNot(HaveOccurred())
-})
+		Expect(err).ToNot(HaveOccurred())
+		return nil
+	},
+	func(_ []byte) {},
+)
 
 var _ = framework.E2EAfterSuite(func() {
+	if skipinboundtags.KubeCluster == nil {
+		return
+	}
 	Expect(skipinboundtags.KubeCluster.DeleteKuma()).To(Succeed())
 	Expect(skipinboundtags.KubeCluster.DismissCluster()).To(Succeed())
 })
